Refactored/removed components and setup layout
